+++ validator/src/main/java/io/avaje/validation/ValidationType.java
@@ -1,12 +1,11 @@
 package io.avaje.validation;
 
 import java.util.Collection;
-import java.util.Set;
 
 
 public interface ValidationType<T> {
 
   void validate(T object) throws ConstraintViolationException;
 
-  void validateAll(Collection<T> collection);
+  //void validateAll(Collection<T> collection);
 }
