+++ validator/src/main/java/io/avaje/validation/core/NoopAnnotationValidator.java
@@ -1,6 +1,6 @@
 package io.avaje.validation.core;
 
-import io.avaje.validation.AnnotationValidationAdapter;
+import io.avaje.validation.adapter.AnnotationValidationAdapter;
 import io.avaje.validation.adapter.ValidationRequest;
 
 public class NoopAnnotationValidator<T> implements AnnotationValidationAdapter<T> {
