+++ validator/src/test/java/io/avaje/validation/core/CustomerValidationAdapter.java
@@ -2,8 +2,8 @@ package io.avaje.validation.core;
 
 import java.util.Map;
 
-import io.avaje.validation.AnnotationValidationAdapter;
-import io.avaje.validation.ValidationAdapter;
+import io.avaje.validation.adapter.AnnotationValidationAdapter;
+import io.avaje.validation.adapter.ValidationAdapter;
 import io.avaje.validation.Validator;
 import io.avaje.validation.adapter.ValidationRequest;
 import jakarta.validation.constraints.AssertTrue;
